style : 패키지 정리 및 네이밍 변경
