+++ src/main/java/baseball/Application.java
@@ -1,7 +1,9 @@
 package baseball;
 
+import baseball.game.BaseballGame;
+
 public class Application {
     public static void main(String[] args) {
-        BaseballGameRunner.run();
+        BaseballGame.run();
     }
 }
