+++ src/main/java/baseball/game/numbers/AnswerNumbers.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.game.numbers;
 
 import camp.nextstep.edu.missionutils.Randoms;
