+++ src/main/java/baseball/io/OutputHandler.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.io;
 
 public class OutputHandler {
